requirements.txt
pygame-ce
pygbag